refactor(amazonq): move getIndentedCode, cleanup telemetry #5765
- move getIndentedCode to textDocumentUtilities
- use cwsprChatInteractionType from aws-toolkit-common
